Return `null` when extracting the value of an empty `Optional` (#2372)
See #2363.
